requirements.txt
"""
Smart grid system requirements
Install with: pip install -r requirements.txt
"""

# Core dependencies
PyYAML>=6.0                    # YAML configuration file parsing
python-dateutil>=2.8.0         # Date/time utilities

# Homomorphic Encryption
tenseal>=0.3.0                # TensorSEAL - CKKS implementation
numpy>=1.20.0                 # Numerical operations

# Testing and development
pytest>=7.0.0                 # Unit testing
pytest-cov>=3.0.0             # Code coverage

# Optional: Performance profiling
psutil>=5.9.0                 # System resource monitoring
